FrameIO-Tools/requirements.txt
click>=8.0.0
python-dotenv>=0.19.0
requests>=2.26.0
rich>=10.0.0 